requirements.txt
APScheduler==3.10.4
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
waitress==3.0.0
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
graphviz==0.20.3
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.4.0
pydot==3.0.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.31
sqlalchemy_schemadisplay==2.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
Werkzeug==3.0.3
